requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
base58==2.1.1
bitarray==2.7.3
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.0.4
colorama==0.4.6
cytoolz==0.12.1
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
ethereum-dasm==0.1.4
evmdasm==0.1.10
frozenlist==1.3.3
h11==0.14.0
hexbytes==0.3.0
httpcore==0.17.0
idna==3.4
ipfshttpclient==0.8.0a2
jsonschema==4.17.3
lru-dict==1.1.8
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
numpy==1.24.2
pandas==2.0.0
parsimonious==0.9.0
protobuf==4.22.3
pycryptodome==3.17
pydantic==2.6.1
pydantic_core==2.16.2
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
regex==2023.3.23
requests==2.28.2
rlp==3.0.0
six==1.16.0
sniffio==1.3.0
tabulate==0.9.0
toolz==0.12.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.15
varint==1.0.2
web3==6.2.0
websockets==11.0.2
yarl==1.8.2
PyYAML~=6.0.1